#22 this is the best "scroll syncing" I could figure out.
the webkit rendering height isn't the same as the avalonedit height, so it runs into issues where its not entirely in sync, but a little bit out. WebKit view can load images in, but the avalonedit view doesn't. Trying to find a solution might be a situation where everybody ("Except Steven: The Chaos Squirrel") loses.

Upgraded to MahApps.Metro 0.4